Fix tick label color
