+++ src/components/Insights/DurationInsight/index.tsx
@@ -61,16 +61,6 @@ const getTickColor = (theme: DefaultTheme) => {
   }
 };
 
-const getTickLabelColor = (theme: DefaultTheme) => {
-  switch (theme.mode) {
-    case "light":
-      return "#4d668a";
-    case "dark":
-    case "dark-jetbrains":
-      return "#dfe1e5";
-  }
-};
-
 const calculateBars = (
   bars: HistogramBarData[],
   chartWidth: number
@@ -136,7 +126,6 @@ export const DurationInsight = (props: DurationInsightProps) => {
   // const config = useContext(ConfigContext);
   const theme = useTheme();
   const tickColor = getTickColor(theme);
-  const tickLabelColor = getTickLabelColor(theme);
   const { observe, width } = useDimensions();
 
   const sortedPercentiles = [...props.insight.percentiles].sort(
